Вынесла вспомогательные функции в отдельный файл
